Add files via upload
Smart Plant Care System (SPCS) is the IoT Capstone Project of IOP490_G4, FPT University Ha Noi. SPCS is a system that allows users to monitor and manage the cultivation of plants, with the main subject of the project being sprouts, in an automated manner. This system uses sensors to measure factors affecting plant growth (temperature, air humidity, soil moisture), along with a microcontroller to send real-time data for monitoring and recording purposes.